+++ @apicize/toolkit/src/controls/viewers/result/result-details-viewer.tsx
@@ -7,29 +7,25 @@ import { observer } from "mobx-react-lite";
 import { useClipboard } from "../../../contexts/clipboard.context";
 import { useWorkspace } from "../../../contexts/workspace.context";
 
-export const ResultRequestViewer = observer((props: {
+export const ResultDetailsViewer = observer((props: {
     requestOrGroupId: string,
     executionResultId: string,
 }) => {
     const workspace = useWorkspace()
     const clipboard = useClipboard()
 
-    const result = workspace.getExecutionResult(props.requestOrGroupId, props.executionResultId)
+    const result = workspace.getExecutionResultDetails(props.requestOrGroupId, props.executionResultId)
 
-    if (result?.type !== 'request') {
-        return null
-    }
+    const request = result?.testingContext
+    if (!request) return null
 
-    const request = result.request
-    if (! request) return null
-
-    const text = beautify.js_beautify(JSON.stringify(request), {})
+    const text = beautify.js_beautify(JSON.stringify(result), {})
     return (
         <Stack sx={{ bottom: 0, overflow: 'hidden', position: 'relative', height: '100%', display: 'flex' }}>
-            <Typography variant='h2' sx={{ marginTop: 0, flexGrow: 0 }} component='div'>Request
+            <Typography variant='h2' sx={{ marginTop: 0, flexGrow: 0 }} component='div'>Details
                 <IconButton
-                    aria-label="copy request to clipboard"
-                    title="Copy Request to Clipboard"
+                    aria-label="copy deatils to clipboard"
+                    title="Copy Details to Clipboard"
                     sx={{ marginLeft: '16px' }}
                     onClick={_ => { if (text) clipboard.writeTextToClipboard(text) }}>
                     <ContentCopyIcon />